requirements.txt
appdirs==1.4.3
astroid==1.6.4
certifi==2018.4.16
chardet==3.0.4
confluent-kafka==0.11.4
coreapi==2.3.3
coreschema==0.0.4
Django~=1.11
django-filter==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
kafka-python==1.4.2
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
packaging==16.8
psycopg2==2.7.4
psycopg2-binary==2.7.4
pygments==2.2.0
pylint==1.9.1
pylint-django==0.11
pylint-plugin-utils==0.2.6
pyparsing==2.2.0
pytz==2017.2
requests==2.18.4
simplejson==3.15.0
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
django-cors-headers==2.3.0
drf-yasg==1.9.1
openpyxl==2.5.4
python-jose==3.0.0
django_excel==0.0.10
